requirements.txt
django-postgrespool==0.3.0
sqlalchemy==1.0.5
amqp==1.4.6
anyjson==0.3.3
astroid==1.3.6
billiard==3.3.0.20
celery==3.1.18
cov-core==1.15.0
coverage==3.7.1
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.8.2
django-bootstrap3==5.4.0
django-extensions==1.5.5
django-sslify==0.2.7
djangorestframework==3.1.3
ElvantoAPI==1.3.3.1
enum34==1.0.4
execnet==1.3.0
fake-factory==0.5.2
gunicorn==19.3.0
hypothesis==1.6.2
hypothesis-fakefactory==0.12.1
ipaddress==1.0.7
ipython==3.2.0
isort==3.9.6
kombu==3.0.26
logilab-common==0.63.2
natsort==4.0.1
oauthlib==0.7.2
opbeat==2.1.1
pep8==1.6.2
pies==2.6.5
pies2overrides==2.6.5
psycopg2==2.6.1
py==1.4.28
pyflakes==0.9.2
PyJWT==1.3.0
pylint==1.4.3
pytest==2.7.1
pytest-cache==1.0
pytest-cov==1.8.1
pytest-django==2.8.0
pytest-flakes==1.0.0
pytest-isort==0.1.0
pytest-pep8==1.0.6
python-social-auth==0.2.10
python3-openid==3.0.5
pytz==2015.4
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
wheel==0.24.0
whitenoise==2.0
